+++ models/events/travel.js
@@ -13,7 +13,7 @@ const { ENUM_EVENT_TYPE, ENUM_ENEMY_STRENGTH,
     ENUM_PERSONALITIES,
     ENUM_SEASONS
 } = require('../../generic/enums')
-const { personalityDealsWith } = require('../../models/personality')
+const { personalityDealsWith } = require('../personality')
 const { getSeason } = require('../../lib/time')
 const bMonster = require('../../build/monster')
 const bCharacter = require('../../build/character')
