+++ models/party.js
@@ -11,7 +11,7 @@ const {
 const { get } = require('../localization')
 
 const { executeCommands } = require('../persistance/commandQueue')
-const { getDwellingByCoordinates } = require('../models/dwelling')
+const { getDwellingByCoordinates } = require('./dwelling')
 const { getStoryEntry } = require('../build/story')
 const { getRoomByCoordinates } = require('../persistance').queries
 const checkForRest = (party) => {
